Atualiza

+++ src/particle.hpp
@@ -7,25 +7,25 @@ using namespace std;
 
 class Particle{
     private:
-        vector<double> pos;
-        vector<double> best_pos;
-        vector<double> velocity;
-        double fitness;
-        double best_fitness;
+        vector<long double> pos;
+        vector<long double> best_pos;
+        vector<long double> velocity;
+        long double fitness;
+        long double best_fitness;
     public:
-        const vector<double>& getVelocity(void) const;
-        void setVelocity(const vector<double> &new_velocity);
+        const vector<long double>& getVelocity(void) const;
+        void setVelocity(const vector<long double> &new_velocity);
 
-        const vector<double>& getPos(void) const;
-        void setPos(const vector<double> &new_pos);
+        const vector<long double>& getPos(void) const;
+        void setPos(const vector<long double> &new_pos);
 
-        const vector<double>& getBestPos(void) const;
-        void setBestPos(const vector<double> &new_best_pos);
+        const vector<long double>& getBestPos(void) const;
+        void setBestPos(const vector<long double> &new_best_pos);
 
-        const double getFitness(void) const;
-        void updateFitness(const double new_fitness);
+        const long double getFitness(void) const;
+        void updateFitness(const long double new_fitness);
 
-        const double getBestFitness(void) const;
+        const long double getBestFitness(void) const;
 
         Particle(const int ndimensions);
         Particle(void);
